Refactor email-client and email-server to mail
